# specs/001-data-layer-postgrest-gitops/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TAVR Data Platform API
  description: |
    REST API for the TAVR Data Infrastructure Platform.

    This API provides access to:
    - Hospital targeting data and scores
    - Data catalog with health monitoring
    - Batch job management and triggers

    The API is powered by PostgREST, which auto-generates endpoints from PostgreSQL views.
  version: 2.0.0
  contact:
    name: Data Platform Team
servers:
  - url: http://localhost:3030
    description: Local development
  - url: https://api.tavr-data.internal
    description: Production (internal)
...
paths:
...
  /jobs:
    get:
      tags:
        - Jobs
      summary: List batch jobs
      description: Returns all configured batch jobs with their current status.
      operationId: getJobs
      parameters:
        - name: is_enabled
          in: query
          description: Filter by enabled status
          schema:
            type: boolean
        - name: last_run_status
          in: query
          description: Filter by last run status
          schema:
            type: string
            enum: [success, failure, running]
      responses:
        '200':
          description: List of batch jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BatchJob'
  /jobs/{job_name}/trigger:
    post:
      tags:
        - Jobs
      summary: Trigger batch job
      description: |
        Manually trigger a batch job to run immediately.
        Creates a Kubernetes Job from the CronJob template.
      operationId: triggerJob
      parameters:
        - name: job_name
          in: path
          required: true
          schema:
            type: string
          example: "fetch-cms-all"
      requestBody:
        description: Optional trigger parameters
        content:
          application/json:
            schema:
              type: object
              properties:
                sources:
                  type: array
                  items:
                    type: string
                  description: Subset of sources to fetch (optional)
                force:
                  type: boolean
                  default: false
                  description: Force run even if recently completed
      responses:
        '202':
          description: Job triggered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobTriggerResponse'
        '400':
          description: Invalid request
        '404':
          description: Job not found
        '409':
          description: Job already running
...
components:
  schemas:
...
    BatchJob:
      type: object
      properties:
        job_id:
          type: integer
        job_name:
          type: string
          example: "fetch-cms-all"
        description:
          type: string
          example: "Fetch all CMS data sources"
        cron_schedule:
          type: string
          example: "0 2 * * 0"
        is_enabled:
          type: boolean
        last_run_at:
          type: string
          format: date-time
        last_run_status:
          type: string
          enum: [success, failure, running]
        last_run_duration_seconds:
          type: integer
        next_scheduled_run:
          type: string
          format: date-time
        source_names:
          type: array
          items:
            type: string
    JobTriggerResponse:
      type: object
      properties:
        job_id:
          type: string
          example: "fetch-cms-all-manual-1705234567"
        status:
          type: string
          example: "created"
        message:
          type: string
          example: "Job triggered successfully"
        k8s_job_name:
          type: string
          example: "fetch-cms-all-manual-1705234567"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated endpoints
security:
  - {} # Anonymous access for read operations
  - bearerAuth: [] # JWT for write operations